+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/core/mod.rs
@@ -1,7 +1,20 @@
-use crate::targets::rust_type::{Definitions, Imports};
+use crate::conversions::Result;
+use crate::targets::rust_type::{Definitions, Imports, ModDef, ModuleName};
+use std::ops::Not;
 
 #[derive(Clone, Debug, Default)]
 pub struct CoreShape {
     pub imports: Imports,
     pub defs: Definitions,
+}
+
+impl CoreShape {
+    pub fn define(self) -> Result<Option<ModDef>> {
+        let def = self.defs.is_empty().not().then(|| ModDef {
+            name: ModuleName::new("core"),
+            imports: self.imports,
+            defs: self.defs,
+        });
+        Ok(def)
+    }
 }
